+++ Changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.4.0.] - 2021/10/12
 ### Changed
 - Removed unused code (config_notif.py)
+### Fixed
+- Pod become unready state
+
 
 ## [2.3.1.] - 2021/06/19
 ### Security