+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.4.0] - 2021/10/12
+### Changed
+- [DCAEGEN2-2939] Removed unused code (config\_notif.py)
+
+
 ## [2.3.1.] - 2021/06/19
 ### Security
 - Fixed SQL injection vulnerability